+++ src/json.cpp
@@ -134,7 +134,7 @@ static void sb_put(SB *sb, const char *bytes, int count)
 
 static void sb_puts(SB *sb, const char *str)
 {
-  sb_put(sb, str, strlen(str));
+  sb_put(sb, str, (int)strlen(str));
 }
 
 static char *sb_finish(SB *sb)
